Added shouldNavigateToLink on NavigatorDelegate
To be able to implement alternative ways of internal links, like presenting it in a separate view, this commit adds a NavigatorDelegate method called `shouldNavigateToLink` that by default returns true - but can be overridden to signal to the caller that we've handled the resource outside of the regular flow.

This can be used to catch things like links to non-linear resources (Impossible to open resource with linear=no #330).